Add new parser option: AllowEmptyExpressions
This new boolean option determines whether the parser should accept input
containing no expressions. If the field is false, and input is "empty",
the parser will reject the input and return an error (with a descriptive message).

Empty interval expressions are problematic since semantically they mean
"match nothing" - and thus would result in Expressions that evaluate
Matches(x) == false for all values of x. This is usually *not* desired.
Additionally, it is very likely for users to accidentally pass empty strings
into the parser. Thus from now on DefaultParseOptions() will set the
field to 'true'.

Note that this commit affects the behavior of ParseExpression():
to revert to old behavior, use ParseExpressionWithOptions().